login logout fixed a close bug
Fixed a bug where server would be prevented from closing by open connectionlistener. Did this by making connectionlistener a daemon thread.

Also added functions for logging accounts in and out and keeping track of which accounts exist and which are logged in. All seems to work properly. moving on to GameServer infrastructure next